refactor: move pipeline services hooks into dedicated file
