+++ src/services/pipeline/useCreatePipeline.ts
@@ -1,0 +1,43 @@
+import { useMutation, useQueryClient } from "react-query";
+import {
+  createPipelineMutation,
+  CreatePipelinePayload,
+  Pipeline,
+} from "@/lib/instill";
+import { constructPipelineRecipeWithDefinition } from "../helper";
+
+const useCreatePipeline = () => {
+  const queryClient = useQueryClient();
+
+  return useMutation(
+    async (payload: CreatePipelinePayload) => {
+      const res = await createPipelineMutation(payload);
+      return Promise.resolve(res);
+    },
+    {
+      onSuccess: async (newPipeline) => {
+        const recipe = await constructPipelineRecipeWithDefinition(
+          newPipeline.recipe
+        );
+
+        const pipeline: Pipeline = {
+          ...newPipeline,
+          recipe: recipe,
+        };
+
+        queryClient.setQueryData<Pipeline>(
+          ["pipelines", newPipeline.id],
+          pipeline
+        );
+
+        queryClient.setQueryData<Pipeline[]>(["pipelines"], (old) =>
+          old
+            ? [...old.filter((e) => e.id !== newPipeline.id), pipeline]
+            : [pipeline]
+        );
+      },
+    }
+  );
+};
+
+export default useCreatePipeline;
